requirements.txt
PyQt6==6.6.1
send2trash==1.8.2
python-dateutil==2.8.2 